@@ -5,7 +5,7 @@ homepage:
 
 environment:
   sdk: '>=2.17.0 <3.0.0'
-  flutter: ">=2.5.0"
+  flutter: ">=3.3.0 <3.9.0"
 
 dependencies:
   flutter:
