@@ -1,54 +1,44 @@
-import 'dart:ui';
-
 import 'package:flutter/material.dart';
-import 'package:linksys_moab/localization/localization_hook.dart';
-import 'package:linksys_moab/page/components/base_components/base_page_view.dart';
-import 'package:linksys_moab/route/_route.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:linksys_app/localization/localization_hook.dart';
+import 'package:linksys_widgets/theme/const/spacing.dart';
+import 'package:linksys_widgets/widgets/_widgets.dart';
 
-import 'package:flutter_gen/gen_l10n/app_localizations.dart';
+import 'package:linksys_widgets/widgets/page/base_page_view.dart';
 
-class NoInternetConnectionModal extends StatelessWidget {
+class NoInternetConnectionModal extends ConsumerWidget {
   const NoInternetConnectionModal({Key? key}) : super(key: key);
 
   @override
-  Widget build(BuildContext context) {
+  Widget build(BuildContext context, WidgetRef ref) {
     return WillPopScope(
-
-      onWillPop: () async { return false; },
-      child: BasePageView.bottomSheetModal(bottomSheet: Container(
-          color: Colors.white,
+      onWillPop: () async {
+        return false;
+      },
+      child: AppPageView.bottomSheetModalBlur(
+        padding: const EdgeInsets.only(),
+        bottomSheet: Container(
+          color: Theme.of(context).colorScheme.primaryContainer,
           height: 240,
           padding: const EdgeInsets.all(24),
           child: Stack(
             children: [
               Padding(
-                padding: const EdgeInsets.all(24),
+                padding: const EdgeInsets.all(Spacing.semiBig),
                 child: Column(
                   mainAxisAlignment: MainAxisAlignment.start,
                   crossAxisAlignment: CrossAxisAlignment.start,
                   children: [
-                    Text(
-                      getAppLocalizations(context).prompt_no_internet_connection,
-                      style: Theme.of(context).textTheme.headline1,
+                    AppText.bodyLarge(
+                      getAppLocalizations(context)
+                          .prompt_no_internet_connection,
                     ),
-                    const SizedBox(
-                      height: 36,
-                    ),
-                    Text(
-                      getAppLocalizations(context).prompt_no_internet_connection_description,
-                      style: Theme.of(context).textTheme.bodyText1,
+                    const AppGap.big(),
+                    AppText.bodyMedium(
+                      getAppLocalizations(context)
+                          .prompt_no_internet_connection_description,
                     )
                   ],
-                ),
-              ),
-              Container(
-                alignment: Alignment.topRight,
-                child: IconButton(
-                  iconSize: 36,
-                  icon: const Icon(Icons.close),
-                  onPressed: () {
-                    NavigationCubit.of(context).pop();
-                  },
                 ),
               ),
             ],
