@@ -1,5 +1,6 @@
-//GENERATED BARREL FILE 
-export 'pref_key.dart'; 
-export 'error_code.dart'; 
-export 'cloud_const.dart'; 
-export 'build_config.dart'; 
+//GENERATED BARREL FILE
+export 'pref_key.dart';
+export 'error_code.dart';
+export 'cloud_const.dart';
+export 'build_config.dart';
+export 'color_const.dart';
