@@ -1,8 +1,55 @@
-enum CloudEnvironment { dev, qa, prod }
+import 'package:collection/collection.dart';
+import 'package:flutter/foundation.dart';
+import 'package:linksys_app/constants/_constants.dart';
+import 'package:linksys_app/core/utils/logger.dart';
+import 'package:shared_preferences/shared_preferences.dart';
 
-class BuildConfig {
-  static const String cloudEnv = String.fromEnvironment('cloud_env', defaultValue: 'qa');
-  static const bool isEnableEnvPicker = bool.fromEnvironment('enable_env_picker', defaultValue: true);
+enum CloudEnvironment {
+  // dev,
+  qa,
+  prod;
+
+  static CloudEnvironment? get(String name) {
+    return values.firstWhereOrNull((element) => element.name == name);
+  }
 }
 
-CloudEnvironment cloudEnvTarget = CloudEnvironment.values.firstWhere((element) => element.name == BuildConfig.cloudEnv);
+enum ForceCommand {
+  local,
+  remote,
+  none;
+
+  static ForceCommand reslove(String type) {
+    logger.d('Force - $type');
+    if (type == 'local') {
+      return ForceCommand.local;
+    } else if (type == 'remote') {
+      return ForceCommand.remote;
+    } else {
+      return ForceCommand.none;
+    }
+  }
+}
+
+class BuildConfig {
+  static const String cloudEnv =
+      String.fromEnvironment('cloud_env', defaultValue: 'qa');
+  static const bool isEnableEnvPicker =
+      bool.fromEnvironment('enable_env_picker', defaultValue: true);
+
+  static ForceCommand forceCommandType = ForceCommand.reslove(
+      const String.fromEnvironment('force', defaultValue: 'none'));
+
+  static load() async {
+    final prefs = await SharedPreferences.getInstance();
+    final envStr = prefs.getString(pCloudEnv);
+    cloudEnvTarget = CloudEnvironment.get(envStr ?? '') ?? cloudEnvTarget;
+  }
+}
+
+bool showDebugPanel = !kReleaseMode && !kIsWeb;
+
+CloudEnvironment cloudEnvTarget = CloudEnvironment.values
+    .firstWhere((element) => element.name == BuildConfig.cloudEnv);
+Map<String, dynamic> get cloudEnvironmentConfig =>
+    kCloudEnvironmentMap[cloudEnvTarget];
