@@ -1,13 +1,21 @@
-
+const kJNAPActionBase = 'http://linksys.com';
+const kJNAPActionPrefix = '$kJNAPActionBase/jnap';
+const kJNAPSession = 'X-JNAP-Session';
+const kJNAPAction = 'X-JNAP-Action';
+const kJNAPAuthorization = 'X-JNAP-Authorization';
+const kJNAPNetworkId = 'X-Linksys-Network-Id';
+//
 const varMqttGroupId = '{group_id}';
 const varMqttNetworkId = '{network_id}';
 
 const keyJnapResult = 'result';
 const keyJnapOutput = 'output';
+const keyJnapSideEffects = 'sideEffects';
 const keyJnapResponses = 'responses';
 const keyJnapError = 'error';
 
 const jnapResultOk = 'OK';
+const jnapResultError = 'Error';
 
 const keyMqttHeader = 'header';
 const keyMqttBody = 'body';
@@ -21,3 +29,4 @@ const mqttRemotePublishTopic = 'groups/$varMqttGroupId/nets/$varMqttNetworkId';
 const mqttRemoteResponseTopic = '$mqttRemotePublishTopic/response';
 
 const userDefinedDeviceLocation = 'userDeviceLocation';
+const userDefinedDeviceName = 'userDeviceName';
