@@ -1,15 +1,24 @@
-// Moab Share Preferences keys, please check is there any data need to be removed when log out
-const moabPrefAppKey = 'MoabApp';
-const moabPrefConfigKey = 'MoabConfigKey';
-const moabPrefCloudCertDataKey = 'MoabCloudCertData';
-const moabPrefCloudPrivateKey = 'MoabCloudPrivateKey';
-const moabPrefCloudPublicKey = 'MoabCloudPublicKey';
-const moabPrefEnableBiometrics = "MoabEnableBiometrics";
-const moabPrefDeviceToken = 'MoabDeviceToken';
-const moabPrefSessionDataKey = 'MoabSessionData';
-const moabPrefLocalCert = 'MoabLocalCert';
-const moabPrefLocalPassword = 'MoabLocalPassword';
-const moabPrefRemoteCaCert = 'MoabRemoteCaCert';
-const moabPrefCloudAccountId = 'MoabCloudAccountId';
-const moabPrefCloudDefaultGroupId = 'MoabCloudDefaultGroupId';
-const moabPrefSelectedNetworkId = 'MoabSelectedNetworkId';
+// Share Preferences keys, please check is there any data need to be removed when log out
+// unsecured
+const pCloudEnv = 'CloudEnvironment';
+const pCurrentSN = 'LinksysCurrentSN';
+const pDeviceToken = 'LinksysDeviceToken';
+const pSmartDeviceId = 'LinksysSmartDeviceId';
+const pSmartDeviceSecret = 'LinksysSmartDeviceSecret';
+const pSmartDeviceVerified = 'LinksysSmartDeviceVerified';
+
+const pSelectedNetworkId = 'LinksysSelectedNetworkId';
+const pShowPushPrompt = 'PushPrompt';
+const pAppFirstLaunch = 'AppFirstLaunch';
+const pWebLog = 'web_log';
+
+const pNotificationLastSeen = 'NotificationLastSeen';
+const pNotifications = 'Notifications';
+
+// secured
+const pSessionToken = 'SessionToken';
+const pSessionTokenTs = 'SessionTokenTimeStamp';
+const pUserPassword = 'UserPassword';
+const pUsername = 'Username';
+const pLocalPassword = 'LocalPassword';
+const pBiometrics = 'Biometrics';
