@@ -1,8 +1,6 @@
-//GENERATED BARREL FILE 
-export 'qr_view.dart'; 
-export 'popup_button.dart'; 
-export 'image_with_badge.dart'; 
-export 'network_check_view.dart'; 
-export 'hidden_password_widget.dart'; 
-export 'no_network_bottom_modal.dart'; 
-export 'password_validity_widget.dart'; 
+//GENERATED BARREL FILE
+export 'qr_view.dart';
+export 'image_with_badge.dart';
+export 'network_check_view.dart';
+export 'hidden_password_widget.dart';
+export 'no_network_bottom_modal.dart';
