@@ -1,6 +1,7 @@
 import 'package:flutter/cupertino.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
 
-class ImageWithBadge extends StatelessWidget {
+class ImageWithBadge extends ConsumerWidget {
   const ImageWithBadge({
     Key? key,
     required this.imagePath,
@@ -8,6 +9,7 @@ class ImageWithBadge extends StatelessWidget {
     required this.imageSize,
     this.badgeSize,
     this.offset = 8,
+    this.fit,
   }) : super(key: key);
 
   final String imagePath;
@@ -15,9 +17,10 @@ class ImageWithBadge extends StatelessWidget {
   final double imageSize;
   final double? badgeSize;
   final double offset;
+  final BoxFit? fit;
 
   @override
-  Widget build(BuildContext context) {
+  Widget build(BuildContext context, WidgetRef ref) {
     return SizedBox(
       width: imageSize + offset,
       height: imageSize + offset,
@@ -30,6 +33,7 @@ class ImageWithBadge extends StatelessWidget {
               imagePath,
               width: imageSize,
               height: imageSize,
+              fit: fit,
             ),
           ),
           if (badgePath != null)
